speed optimization -O3 and platform.txt, smaller data types make the Arduino faster

+++ CodingSwitch.h
@@ -3,7 +3,7 @@ The MIDI output channel 1..16 (or 0..15) is selected with a binary coding switch
 5 resistors map the digital 4 bit values to analog voltages. 
 */
 
-const int n_values = 10;
+const byte n_values = 10;
 const int values[n_values] = {0,510,311,602,173,558,401,636,90,534};
 const int delta = 10;
 
@@ -19,7 +19,7 @@ byte readCodingSwitchValue(int pin) {
   Serial.print(value);
   Serial.print(" -> ");
   #endif
-  for (int i = 0 ; i < n_values; i++) {
+  for (byte i = 0 ; i < n_values; i++) {
     if (value <= values[i] + delta && value >= values[i] - delta) {
       #ifdef DEBUG_CODING_SWITCH
       Serial.println(i);
